rename init to setGlobalHandler
